各ValueObjectの明示的型変換を暗黙的型変換に置換

+++ WodiLib/WodiLib.Test/Event/CharaMoveCommand/ValueObject/CharaMoveCommandValueTest.cs
@@ -99,7 +99,7 @@ namespace WodiLib.Test.Event.CharaMoveCommand
             var errorOccured = false;
             try
             {
-                castValue = (int) instance;
+                castValue =  instance;
             }
             catch (Exception ex)
             {
