+++ WodiLib/WodiLib.Test/Map/ValueObject/ShadowGraphicIdTest.cs
@@ -80,7 +80,7 @@ namespace WodiLib.Test.Map
             var errorOccured = false;
             try
             {
-                castValue = (int) instance;
+                castValue =  instance;
             }
             catch (Exception ex)
             {
