+++ WodiLib/WodiLib.Test/Map/ValueObject/TileSetIdTest.cs
@@ -84,7 +84,7 @@ namespace WodiLib.Test.Map
             var errorOccured = false;
             try
             {
-                castValue = (int) instance;
+                castValue =  instance;
             }
             catch (Exception ex)
             {
